+++ cli/tests/test_split_command.rs
@@ -18,7 +18,7 @@ use std::path::PathBuf;
 use crate::common::TestEnvironment;
 
 fn get_log_output(test_env: &TestEnvironment, cwd: &Path) -> String {
-    let template = r#"separate(" ", change_id.short(), empty, description, local_bookmarks)"#;
+    let template = r#"separate(" ", change_id.short(), empty, local_bookmarks, description)"#;
     test_env.jj_cmd_success(cwd, &["log", "-T", template])
 }
 
@@ -226,8 +226,9 @@ fn test_split_with_default_description() {
     std::fs::write(workspace_path.join("file1"), "foo\n").unwrap();
     std::fs::write(workspace_path.join("file2"), "bar\n").unwrap();
 
-    // Create a bookmark pointing to the commit. It will be moved to the second
-    // commit after the split.
+    // Create a bookmark pointing to the target commit. It will move to the
+    // first commit created by the split, which will retain the change id of the
+    // target commit.
     test_env.jj_cmd_ok(&workspace_path, &["bookmark", "create", "test_bookmark"]);
 
     let edit_script = test_env.set_up_fake_editor();
@@ -239,10 +240,10 @@ fn test_split_with_default_description() {
     let (stdout, stderr) = test_env.jj_cmd_ok(&workspace_path, &["split", "file1"]);
     insta::assert_snapshot!(stdout, @"");
     insta::assert_snapshot!(stderr, @r###"
-    First part: qpvuntsm 48018df6 TESTED=TODO
-    Second part: kkmpptxz 350b4c13 test_bookmark | (no description set)
-    Working copy now at: kkmpptxz 350b4c13 test_bookmark | (no description set)
-    Parent commit      : qpvuntsm 48018df6 TESTED=TODO
+    First part: qpvuntsm 48018df6 test_bookmark | TESTED=TODO
+    Second part: kkmpptxz 350b4c13 (no description set)
+    Working copy now at: kkmpptxz 350b4c13 (no description set)
+    Parent commit      : qpvuntsm 48018df6 test_bookmark | TESTED=TODO
     "###);
 
     // Since the commit being split has no description, the user will only be
@@ -263,8 +264,8 @@ fn test_split_with_default_description() {
     "###);
     assert!(!test_env.env_root().join("editor2").exists());
     insta::assert_snapshot!(get_log_output(&test_env, &workspace_path), @r###"
-    @  kkmpptxzrspx false test_bookmark
-    ○  qpvuntsmwlqt false TESTED=TODO
+    @  kkmpptxzrspx false
+    ○  qpvuntsmwlqt false test_bookmark TESTED=TODO
     ◆  zzzzzzzzzzzz true
     "###);
 }
@@ -334,8 +335,9 @@ fn test_split_siblings_no_descendants() {
     std::fs::write(workspace_path.join("file1"), "foo\n").unwrap();
     std::fs::write(workspace_path.join("file2"), "bar\n").unwrap();
 
-    // Create a bookmark pointing to the commit. It will be moved to the second
-    // commit after the split.
+    // Create a bookmark pointing to the target commit. It will move to the
+    // first commit created by the split, which will retain the change id of the
+    // target commit.
     test_env.jj_cmd_ok(&workspace_path, &["bookmark", "create", "test_bookmark"]);
     insta::assert_snapshot!(get_log_output(&test_env, &workspace_path), @r###"
     @  qpvuntsmwlqt false test_bookmark
@@ -351,15 +353,15 @@ fn test_split_siblings_no_descendants() {
     let (stdout, stderr) = test_env.jj_cmd_ok(&workspace_path, &["split", "--parallel", "file1"]);
     insta::assert_snapshot!(stdout, @"");
     insta::assert_snapshot!(stderr, @r###"
-    First part: qpvuntsm 0dced07a TESTED=TODO
-    Second part: zsuskuln 0473f014 test_bookmark | (no description set)
-    Working copy now at: zsuskuln 0473f014 test_bookmark | (no description set)
+    First part: qpvuntsm 0dced07a test_bookmark | TESTED=TODO
+    Second part: zsuskuln 0473f014 (no description set)
+    Working copy now at: zsuskuln 0473f014 (no description set)
     Parent commit      : zzzzzzzz 00000000 (empty) (no description set)
     Added 0 files, modified 0 files, removed 1 files
     "###);
     insta::assert_snapshot!(get_log_output(&test_env, &workspace_path), @r###"
-    @  zsuskulnrvyr false test_bookmark
-    │ ○  qpvuntsmwlqt false TESTED=TODO
+    @  zsuskulnrvyr false
+    │ ○  qpvuntsmwlqt false test_bookmark TESTED=TODO
     ├─╯
     ◆  zzzzzzzzzzzz true
     "###);
@@ -702,7 +704,6 @@ fn test_split_with_multiple_workspaces_same_working_copy() {
         &main_path,
         &["workspace", "add", "--name", "second", "../secondary"],
     );
-    // Change the working copy in the second workspace.
     test_env.jj_cmd_ok(
         &secondary_path,
         &["edit", "-r", "description(first-commit)"],